feat: restructure repository for a standardized file system
